Shiny App with 'babynames" package (v1)
This is initial commit and has the basic app structure (UI and Server functions). Using  one reactive expression to render a ggplot. When a name is provided in the text box and year range is selected via slider, the plot updates accordingly showing the number of occurences of the selected name in a specific year via the line graph