Using observe (WIP)
This code is not quite working yet.

Goal to add a observe call that would calculate the number of rows/observations returned when a specific name/multiple names are provided in a specific year range. Messages are printed to the console saying if a name is popular vs not. If the number of observations are >=30, I am making an assumption that the name is quite popular and observe prints the message to console

Known issues and WIP

1. When I select multiple names via selectInput, the popularity is calculated only for the first name in the box

"Warning: Error in : Problem with `filter()` input `..3`.
ℹ Input `..3` is `Name == input$name`.
x Input `..3` must be of size 1924665 or 1, not size 0.
  79: <Anonymous>"

2. I ideally would like to use the Count column value to assess the popularity instead of number of rows. Hopefully will do that in the next commit